docs: Update documentation and rename file
